chore(actor) Rename RpcActor -> Actor
With the Actor Django model removed we can use the term `Actor` for what
is now the only actor in the application. Once this lands I'll update
all the usage in both sentry & getsentry.
